==> prep/fis/2_noaa_spatial_fish_catch.Rmd <==
---
title: "Spatializing fish catch in the Northeast"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script calculates catch by OHI region using landings data provided by NOAA via data request.

# Data

**Downloaded**: July, 2019 (sent via email after data request)

**Description**:  Commercial fish landings by statistical area

**Time range**: 1996-2017. Data provided annually

**Format**:  Excel spreadsheet

***

# Setup

```{r setup, message = FALSE, warning = FALSE}
knitr::opts_chunk$set(message = FALSE, warning = FALSE, fig.path = "figs/")

library(tidyverse)
library(readxl)
library(sf)
#install.packages("viridis")
library(viridis)
library(janitor)

source("~/github/ne-prep/src/R/common.R")
```


# Data cleaning

Cleaning the raw data a bit by fixing column names and turning `stat_area` numeric. 
```{r}
#loading data
raw <- read_excel(file.path(dir_anx, "_raw_data/NOAA_NMFS/catch_by_stat_area/Afflerbach_UCSB_Landings by Stat Area_JUL 2019 Updated.xlsx"))

#cleaning up a bit
clean <- raw %>%
  rename(year = YEAR,
         stat_area = `STAT\r\nAREA`, 
         species = SPECIES,
         pounds = `LBS LANDED \r\n(HAIL WT)`,
         stock_id = `STOCK ID`,
         stock = `STOCK NAME`) %>%
  mutate(stat_area = as.numeric(stat_area))

#preview of what data looks like
DT::datatable(head(clean, n = 20))
```

There are a total of `r length(unique(clean$species))` reported species, which include groupings such as "Confidential Species" or "Clam, species not specified". Some of these species, such as Atlantic Cod, have multiple stocks. There are `r clean %>% filter(!is.na(stock_id)) %>% distinct(stock_id) %>% nrow()` stock ids, and `r clean %>% filter(is.na(stock_id),!is.na(species)) %>% distinct(species) %>% nrow()` species without distinct stock ids. If we combine these, and remove "Confidential species" and anything identified as "other" we have a total of `r clean %>% filter(is.na(stock_id), !is.na(species)) %>% distinct(species) %>%  filter(!str_detect(species, "OTHER"), !str_detect(species, "CONFIDENTIAL")) %>% nrow()` stocks.


# Spatial data

Since the data is provided by statistical landing area, we can use this information to infer what OHI region's encompass or overlap with these areas. We have downloaded the shapefile for Statistical Areas from [this public FTP NOAA site](ftp://ftp.nefsc.noaa.gov/pub/gis/).

## Statistical areas

Load in the statistical areas and add area of each polygon as a column.

```{r statistical_areas}
stat_shp <- sf::read_sf(file.path(dir_anx, "spatial/Statistical_Areas_2010_withNames.shp")) %>%
  st_set_crs(p4s_nad83) %>% #set the CRS
  st_transform(crs = crs(rgns))  #transform to our NE specific CRS

stat_shp$stat_area <- st_area(stat_shp) #add area as column


ggplot(stat_shp) +
  geom_sf() +
  theme_bw() +
  labs(title = "Statistical areas") +
  theme(legend.position = "none")
```

We overlay statistical areas with our regions to select just the statistical areas that overlap with our regions.

```{r ohi_stat_areas}
ohi_stat_areas          <- st_intersection(rgns, stat_shp) #intersects statistical areas with OHI regions
ohi_stat_areas$ohi_area <- st_area(ohi_stat_areas) #calculate area of each overlapped polygon

ggplot(ohi_stat_areas) +
  geom_sf(aes(fill = rgn_name)) +
  geom_sf(data = ne_states, fill = "beige") +
  theme_bw() +
  labs(title = "Statistical areas overlapped with OHI regions") 
```

Calculate proportion of each statistical area in our OHI regions. For statistical areas that overlap with OHI regions, we can use proportional area overlap to adjust catch. We assume that catch is evenly distributed across each statistical area.

```{r prop_stat_areas_in_ohi_rgns}
calc_prop_area <- ohi_stat_areas %>%
  group_by(Id) %>%
  mutate(ohi_rgn_prop_area = ohi_area/stat_area) #this column tells us how much of each OHI sub-region falls within the statistical area in our region

ggplot(calc_prop_area) +
  geom_sf(aes(fill = ohi_rgn_prop_area)) +
  geom_sf(data = ne_states, fill = "beige") +
  theme_bw() +
  labs(title = "Proportion of each \nstatistical area in OHI region") 
```

# Catch per OHI region

Now we calculate the total catch per species and year for each of the OHI regions, and for the entire Northeast.

First let's filter the catch data to just the statistical areas in our region. We don't care about the catch outside of these statistical areas.

```{r}
rgn_catch <- clean %>%
  filter(stat_area %in% ohi_stat_areas$Id) %>%
  left_join(calc_prop_area, by = c("stat_area" = "Id")) %>%
  mutate(catch = pounds*ohi_rgn_prop_area) %>% #adjusting catch by the proportional area with overlap
  select(-area_km2, -FULL_NAME, -SHORT_NAME, -stat_area.y, -ohi_area, -NAFODIV) %>%
  group_by(species, stock_id, stock, rgn_id, year, rgn_name) %>%
  summarize(catch = sum(catch)) %>%
  ungroup() %>%
  mutate(display_name = ifelse(is.na(stock_id), species, stock_id)) 

#get catch for whole OHI Northeast region
ne_catch <- rgn_catch %>%
  group_by(year, display_name, species, stock_id, stock) %>%
  summarize(catch = sum(catch, na.rm = T)) %>%
  mutate(rgn_name = "Northeast",
         rgn_id = 12)
  
region_catch <- bind_rows(rgn_catch, ne_catch)

head(region_catch, n = 20)
```


# Replace lobster catch

The catch figures for American lobster in this spatial NMFS dataset are significantly lower than reported landings in scientific stock assessments. Given the importance of this species both culturally and commercially, the landings figures are replaced using Lobster data from another NMFS dataset called the Data Matching and Imputation System (DMIS) database. These come with statistical areas so they can matchup with the VTR database. The VTR database is inaccurate when it comes to lobster because vessels with **only a lobster permit** are not required to submit a vessel trip report. The DMIS database builds off the VTR data to estimate landings by area. 


Load the DMIS data and calculate catch per OHI region
```{r}
lobster <- read_excel(file.path(dir_anx, "_raw_data/NOAA_NMFS/JAfflerbach_Landings by Stat Area DMIS_JAN 2020.xlsx")) %>% 
  clean_names() %>%
  filter(species == "LOBSTER, AMERICAN") %>%
  group_by(stat_area, year, species, stock_id, stock_name) %>%
  summarize(catch = sum(lbs_landed)) %>%
  ungroup() %>%
  mutate(stat_area = as.numeric(stat_area)) %>%
  left_join(calc_prop_area, by = c("stat_area" = "Id")) %>%
  mutate(catch = catch*ohi_rgn_prop_area) %>% #adjusting catch by the proportional area with overlap
  select(-area_km2, -FULL_NAME, -SHORT_NAME, -stat_area.y, -ohi_area, -NAFODIV) %>%
  group_by(year, species, rgn_name, rgn_id) %>%
  summarize(catch = sum(catch)) %>%
  ungroup() %>%
  mutate(display_name = ifelse(rgn_id %in% c(4, 5, 10, 11, 8, 1), "AMERICAN LOBSTER SOUTHERN NE", "AMERICAN LOBSTER GULF OF MAINE"),
         stock_id = ifelse(rgn_id %in% c(4, 5, 10, 11, 8, 1), "LOBSTERSNE", "LOBSTERGOM"),
         stock = ifelse(rgn_id %in% c(4, 5, 10, 11, 8, 1), "SNE Lobster", "GOM/GB Lobster")) %>% #differentiate stocks based on north/south cape cod
  filter(!is.na(rgn_id))
```

Create a dataframe of catch for each lobster stock for entire Northeast (region 12)
```{r}
ne_lobster <- lobster %>%
  group_by(species, stock_id, stock, year, display_name) %>%
  summarize(catch = sum(catch)) %>%
  ungroup() %>%
  mutate(rgn_id = 12, rgn_name = "Northeast")
```

Calculate the proportion of catch for the whole region between these two stocks for 2007 and we are going to apply that proportion to the catch for 2005 and 2006.
```{r}
ne_lob_prop <- ne_lobster %>%
  group_by(year, rgn_id, rgn_name) %>%
  mutate(tot_catch = sum(catch)) %>%
  ungroup() %>%
  mutate(prop_catch = catch/tot_catch) %>%
  filter(year == 2007) %>%
  select(species, stock_id, stock, display_name, prop_catch)
```

Grab data from `region_catch` for lobster pre 2007 and assign the Northeast region (region 12) the catch among the two stocks based on the proportional catch just calculated.

```{r}
rgn_12_lobster <- region_catch %>%
  filter(rgn_id == 12,
         species == "LOBSTER, AMERICAN",
         year < 2007) %>%
  left_join(ne_lob_prop, by = "species") %>%
  select(rgn_id, year, rgn_name, species, stock_id = stock_id.y, stock = stock.y, display_name = display_name.y, catch, prop_catch) %>%
  mutate(catch2 = catch*prop_catch) %>%
  select(-catch, -prop_catch) %>%
  rename(catch = catch2)
```


Create final catch dataframe including the updated lobster data

```{r}
rgn_catch_plus_lob <- region_catch %>% 
  mutate(remove = case_when(
    species == "LOBSTER, AMERICAN" & year > 2006 ~ "remove",
    species == "LOBSTER, AMERICAN" & rgn_id == 12 ~ "remove",
    TRUE ~ "keep")) %>%
  filter(remove == "keep") %>% #we want to replace the 2007 onwards lobster
  select(-remove) %>%
  bind_rows(lobster) %>%
  bind_rows(ne_lobster) %>%
  bind_rows(rgn_12_lobster) %>%
  mutate(stock_id = case_when(
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(4, 5, 10, 11, 8, 1) ~ "LOBSTERSNE",
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(2, 3, 6, 7, 9) ~ "LOBSTERGOM",
            TRUE ~ as.character(stock_id)),
         stock = case_when(
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(4, 5, 10, 11, 8, 1) ~ "SNE Lobster",
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(2, 3, 6, 7, 9) ~ "GOM/GB Lobster",
            TRUE ~ as.character(stock)),
         display_name = case_when(
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(4, 5, 10, 11, 8, 1) ~ "AMERICAN LOBSTER SOUTHERN NE",
            species == "LOBSTER, AMERICAN" & rgn_id %in% c(2, 3, 6, 7, 9) ~ "AMERICAN LOBSTER GULF OF MAINE",
            TRUE ~ as.character(display_name))
         )
```

```{r}
lob <- rgn_catch_plus_lob %>%  filter(species == "LOBSTER, AMERICAN")

ggplot(lob, aes(x = year, y = catch, color = display_name)) +
  geom_line() +
  theme_bw() +
  facet_wrap(~rgn_name, scales = "free_y")
```


```{r}
write_csv(rgn_catch_plus_lob, "data/nmfs_spatial_catch_by_ohi_rgn.csv") #this file is used in fis_species_lookuptable.Rmd to create the species lookup table
```


## Total catch over time

```{r total_northeast_catch}
total_catch <- rgn_catch_plus_lob %>%
  filter(rgn_id == 12) %>%
  group_by(year) %>%
  summarize(catch = sum(catch))

ggplot(total_catch, aes(x = year, y = catch)) +
  geom_area() +
  labs(x = "",
       y = "Pounds",
       title= "Total catch in the Northeast") +
  theme_bw()
```

## Visualize catch by region

```{r raw_catch_by_region}
ggplot(rgn_catch_plus_lob, aes(x = year, y = catch, fill = display_name)) +
  facet_wrap(~rgn_name, scales = "free_y") +
  geom_bar(stat = "identity") +
  theme_bw() +
  theme(legend.position = 'none',
        axis.text = element_text(size = 6),
        axis.text.x = element_text(angle = 45))
```

# Gapfill

The data shared with us includes records of 0 catch. But there is still missing data. As an example, let's look at ALEWIFE.

```{r}
ale <- clean %>%
  filter(species == "ALEWIFE")

unique(ale$year)
```

Ok clearly we are missing data for 2001, 2002, 04-05, 2008-11. We don't know if these are 0's or missing data. We need to **gapfill** this missing data. When a species/state combination has missing data for a year, we can not assume it has a catch of 0. Since we calculate a rolling average of catch, NAs will remain as NA's and the average will rely on just one or two years of catch. This is done to account for any wild fluctuations in catch year to year.

```{r gapfilled_catch_by_region}
gf_data <- rgn_catch_plus_lob %>%
  group_by(rgn_id, rgn_name, species, stock, stock_id, display_name) %>%
  complete(year = 1998:2017) %>%
  arrange(year) %>%
  mutate(mean_catch = zoo::rollapply(catch, 3, mean, fill = NA, align = 'right')) %>% ## create a new column `mean_catch` with rolling mean of 3 yrs
  filter(year > 2004) %>%
  select(year, rgn_id, rgn_name, species, stock, stock_id, mean_catch, display_name) %>%
  ungroup()

ggplot(gf_data, aes(x = year, y = mean_catch, fill = display_name)) +
  geom_bar(stat = "identity") +
  facet_wrap(~rgn_name, scales = "free_y") +
  theme_bw() +
  theme(legend.text = element_text(size = 6),
        legend.position = "below",
        axis.text = element_text(size = 6),
        axis.text.x = element_text(angle = 45))
```


Let's look at total regional catch for each species (not stock)

```{r catch_by_species}
#calculate total regional catch per species
species_catch <- gf_data %>%
  group_by(species, year) %>%
  summarize(sp_catch = sum(mean_catch, na.rm=T)) %>%
  ungroup() %>%
  group_by(year) %>%
  mutate(yr_catch = sum(sp_catch),
         catch_prop = sp_catch/yr_catch) %>%
  ungroup() %>%
  filter(year > 2004) 

ggplot(species_catch %>% filter(catch_prop > 0.01), aes(x = year, y = catch_prop, fill = species)) +
  geom_bar(stat = "identity") +
  theme_bw() +
  theme(legend.text = element_text(size = 6))

```

Clearly atlantic herring is making up the majority of catch! Atlantic herring is primarily a bait fishery, so we need to account for that since this goal is only measuring catch meant for human consumption. We adjust for this below.

# Remove non-food catch

Some of these species are harvested for food as well as other markets like pet food or bait. We want to make sure this goal captures catch meant for human consumption. We have data from NOAA that identifies the amount of catch per species, state and year meant for food, bait, and other markets. This data was cleaned in `prop_catch_food_bait.Rmd`.

```{r tot_catch_as_bait}
prop_data <- read_csv("data/fish_catch_food_prop_rgn.csv")

ggplot(prop_data %>% filter(market == "BAIT"), aes(x = year, y = pounds_live_by_market, color = species)) +
  geom_line() +
  theme_bw() +
  theme(legend.title = element_blank()) +
  labs(x = "",
       y = "Catch (pounds)",
       title = "Total amount of species catch sold as bait \n(not for direct human consumption)")


```

Removing the amount of catch that goes to non-human consumption


```{r remove_bait_proportion}
toolbox_data <- gf_data %>%
  left_join(prop_data) %>%
  mutate(prop = ifelse(is.na(prop),1,prop),
         mean_catch_times_prop = mean_catch*prop) %>%
  filter(!market %in% c("BAIT", "NO MARKET", "CANNED PET FOOD"), #remove bait, pet food and no market records.
         !is.na(mean_catch)) %>%  #remove records with no catch (don't need them)
  select(-market, -pounds_live_by_market, -total_pounds_live, -prop)
```



```{r total_catch_by_ohi_region}
tot <- toolbox_data %>%
  filter(rgn_id != 12) %>%
  group_by(year, rgn_name) %>%
  summarize(catch = sum(mean_catch_times_prop))

ggplot(tot, aes(x = year, y = catch, fill = rgn_name)) +
  geom_area() +
  scale_fill_viridis_d() +
  theme_bw() +
  theme(legend.title =  element_blank()) +
  labs(x = "",
       y = "Pounds",
       title = "Total catch in pounds meant for human consumption, by OHI region")
```

## Clean up display names

The raw data from NMFS comes with species names displayed like "SCALLOP, SEA" instead of Sea Scallop. This little fix changes the species names to a more user-friendly version.

```{r clean_up_names}
names <- read_csv("data/display_names.csv")

toolbox_data <- toolbox_data %>%
  left_join(names)
```


# Save to toolbox

```{r save_to_toolbox}
# save to toolbox
write.csv(toolbox_data, file = file.path(dir_calc, "layers/fis_meancatch.csv"))
```
