==> prep/fis/6_stock_scores.Rmd <==
---
title: "Northeast fisheries stock scores: Combining NOAA and RAM data"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/ne-prep/src/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

# Summary

This script creates the stock score layer for the toolbox by combining both NMFS and RAM data.

# Setup
``` {r setup, echo = FALSE, message = FALSE, warning = FALSE, results = 'hide'}

knitr::opts_chunk$set(message = FALSE, warning = FALSE)

source('~/github/ne-prep/src/R/common.R')  ### an OHINE specific version of common.R
```

## Load data

NMFS/NOAA stock scores. We also add `rgn_id` since the toolbox needs that column to run.
```{r}
nmfs_ss <- read.csv("data/nmfs_stock_scores.csv") %>% 
  separate(stock, into = c("species", "location"), sep = " - ") %>%
  mutate(source = "NMFS") %>%
  select(year, source, species, location, score)
```

RAM
```{r}
ram_ss <- read.csv("data/ram_stock_scores.csv") %>% 
  separate(stocklong, into = c("species", "location", "location2", "location3", "location4", "location5"), 
           sep = "(?<=.)(?=[[:upper:]])") %>%
  mutate(source = "RAM",
         location = paste(trimws(location), location2, location3, location4, location5),
         location = trimws(str_replace_all(location, "NA", "")),
         location = str_replace_all(location, "  ", " "),
         species = trimws(species)) %>%
  filter(year>2004) %>%
  select(year, source, species, location, score) 
```

# Combine

We can combine these two (RAM and NMFS) now. Then we need to assign region IDs to each stock for the toolbox. We can do this by using the fisheries lookup table

```{r}
#fisheries lookup table gives us a list of each species with stock assessment plus info on where they are
lookup_table <- read_csv("data/assessed_species_lookup_table.csv") %>%
  rename(species  = stock_assessment_species_name,
         location = stock_assessment_species_location)

ss_toolbox <- nmfs_ss %>%
  bind_rows(ram_ss) %>%
  mutate(species = tolower(species)) %>%
  left_join(lookup_table, by = c("source", "species", "location")) %>%
  select(year, source, nmfs_original_species, location, score, rgn_id, stockid)
```


# Save to toolbox
```{r}
write_csv(ss_toolbox, file.path(dir_calc, "layers/fis_stockscores.csv"))
```

# Visualize

I want to see stock scores combined with catch proportion.

```{r}
catch <- read_csv(file.path(dir_calc, "layers/fis_meancatch.csv"))

#combine catch and stock scores
catch_ss <- ss_toolbox %>%
  left_join(catch, by = c("year", "rgn_id", "nmfs_original_species" = "species", "stockid" = "stock_id")) %>%
  filter(!is.na(mean_catch)) %>%
  group_by(year, rgn_id) %>%
  mutate(yr_rgn_catch = sum(mean_catch_times_prop),
         catch_prop = mean_catch_times_prop/yr_rgn_catch) %>%
  ungroup() %>%
  mutate(display_name = ifelse(is.na(stockid), nmfs_original_species, stockid))
```

What proportion of total catch is assessed?

```{r}
total_catch <- read_csv("data/nmfs_spatial_catch_by_ohi_rgn.csv") %>%
  filter(rgn_id == 12) %>%
  group_by(year) %>%
  summarize(total_catch = sum(catch))

assessed_catch <- read_csv("data/nmfs_spatial_catch_by_ohi_rgn.csv") %>%
  filter(rgn_id == 12,
         species %in% lookup_table$nmfs_original_species) %>%
  group_by(year) %>%
  summarize(assessed_catch = sum(catch)) %>%
  left_join(total_catch) %>%
  mutate(prop_assessed = assessed_catch/total_catch)

ggplot(assessed_catch) +
  geom_col(aes(x = year, y = prop_assessed)) +
  theme_minimal() +
  labs(x = "", y = 'Proportion of catch assessed')
  
```


Lines are weighted based on the catch_prop

```{r}
ggplot(catch_ss %>% filter(catch_prop > 0.1), aes(x = year, y = score, color = fixed_name, size=catch_prop)) +
  geom_line() +
  theme_bw() +
  facet_wrap(~rgn_id) +
  theme(legend.position = "bottom")
```

Atlantic herring stock score drops very low due to very high B/Bmsy in 2012 ([see kobe plots here]( https://ohi-northeast.github.io/ne-prep/prep/fis/stock_status_noaa.html#32_visualize))
